Renamed MwpApi into ManagementApi and moved it to management package
